Add files via upload
This is a cricket game where both the player and the computer select a number from 1 to 6. If the chosen numbers match, the player's wicket is taken. In this game, the player aims to achieve a target score within a specified number of balls and wickets. The computer employs a machine learning algorithm to predict the player's number, introducing a high level of competition and excitement.
The target score falls within the range of 30 to 100, and the number of balls available for play is between 12 to 36. To make the game more challenging and engaging, the target economy must exceed 12. Additionally, the game features immersive elements such as music and sound effects, adding to the overall excitement.